Add feature deltas after update
When an update query is executed quit will apply all changes for each file that
is affected. Therefor a dictionary containg triples, graphs and actions
(insert, delete) is generated. Also the file content is held in memory, so
additions and deletions can be applied without the need of a serialization for
each named graph (current behaviour).

TODO:
Add/Remove quads in store instead of executing the update query again.