Add modify test including WITH
Using WITH will make use of default graph in update.py when generating the
delta.
This commit will use the graph IRI given after WITH statement to serialize
a correct delta that will be applied to the file system.

Add tests combinating WITH, USING and USING NAMED over DELETE, INSERT and
DELETE + INSERT