Update data structure of graphdiff delta
Since QuitStore was changed to accept multioperational updates in
update.py, the data structure of the produced delta changed.
Due to missing tests of feature provenance we didn't recognized that the
creation of provenance triples wasn't working with the "new" delta.
This commit changes the data structure and the process of creating
provenance triples.

Fix #145.